refactoring plus new feature
* refactoring. Refactored the way the Ui interacts with
the service and the other way round. The service callbacks to ui
were useless therefore we switched to a remote service for broader
possibilities. Therefore, the communication with the
service is in one direction, Ui to service.
* feature. Added a widget for the application.
* Bugfixing. Several bugfixes were added, including the zoom in
once the app comes back to foreround.
* Bugfixing. Fixed bugs related to controls. Not all the time
the Ui of the controls was up to date.